added an prototype instance for FireBot API
This script is a Financial Analyzer, which traders can use to make investment & trading decisions. Still need to make it work with a cryptocurrency data set, where I can use the same API as the ticker info instance, but I think CoinGecko API would be something better to use. SeanCode