Added skeleton blockchain
This is the first instance of me building a rather simplistic blockchain or a skeleton structure of it. This is script is called if you run mine.py and performs
a basic p2p transaction as a crypto wallet or a miner node would do. Please note this is very simplistic, and just something to build upon as I go along